Move node related deserialize(r) into its own module
